+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.1.0
+- Replaced old Bosch Indego authentication by Bosch SingleKey ID #171
+- Changed to new Bosch Indego API server #171
+
+## 2.0.33
+- Added storage of mowers list during login
+
+## 2.0.30
+- Adjusted python code for some changes in the Bosch API
+
 ## 2.0.29
 - Added new mower model S+ 700 gen 2 #120